pyproject.toml
[project]
name = "brownfield-cartographer"
version = "0.1.0"
description = "Multi-agent codebase intelligence system that generates a living architectural map."
readme = "README.md"
requires-python = ">=3.12"
authors = [{ name = "10 Academy" }]

dependencies = [
  "pydantic>=2.8,<3",
  "networkx>=3.3,<4",
  "typer>=0.24,<1",
  "rich>=13.7,<14",
  # Intended (Phase 3): tree-sitter-based parsing. In restricted-network environments,
  # we fall back to stdlib AST parsing and keep the analyzer interface stable.
  # "tree-sitter>=0.22",
  # "tree-sitter-languages>=1.10",
  "tree-sitter==0.20.4",
  "tree-sitter-languages>=1.10.2",
  "numpy>=2.4.3",
  "sqlglot>=20.0,<25",
  "pyyaml>=6.0",
  "python-dotenv>=1.0,<2",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0,<10",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.uv]
dev-dependencies = ["pytest>=9.0,<10"]

